【吴】能耗分析 十一、根据灯具唯一编码统计单个灯具的电压、能耗信息

+++ ctldb/src/main/java/com/nnlightctl/jdbcdao/impl/EnergyStatisticDaoImpl.java
@@ -259,7 +259,7 @@ public class EnergyStatisticDaoImpl implements EnergyStatisticDao {
             public ListEleboxEnergyStatisticView mapRow(ResultSet resultSet, int i) throws SQLException {
                 ListEleboxEnergyStatisticView listEleboxEnergyStatisticView = new ListEleboxEnergyStatisticView();
                 listEleboxEnergyStatisticView.setUid(resultSet.getString("uid"));
-                listEleboxEnergyStatisticView.setTotal(resultSet.getLong("total"));
+                listEleboxEnergyStatisticView.setTotal(resultSet.getBigDecimal("total"));
                 return listEleboxEnergyStatisticView;
             }
         });
